- Introduce very simple `Scheduler` component - Split components into separate files - Modify `LlmDecoder` to use new flow
